+++ Assets/WorldGeneration/_Scripts/Spawning/TerrainAssets/Interactables.cs
@@ -1,9 +1,0 @@
-using UnityEngine;
-
-namespace WorldGeneration._Scripts.Spawning.TerrainAssets
-{
-    public interface Interactables
-    {
-        public bool Interact(GameObject interactingObject);
-    }
-}